requirements.txt
attrdict
gpustat
torchtext
scipy
ipywidgets
ipdb
gensim
flask
networkx
celluloid
gensim
nltk
pydot
matplotlib
numpy
pandas
scikit-learn
seaborn
tqdm==4.64.0
transformers
wandb==0.12.11
xgboost
protobuf==3.20.1
munch
einops